techdebt: delete IVirtualHeap; session interfaces move to the SDK
IVirtualHeap was a vestige of a god-class since decomposed into
IRuntimeSession -> ISessionSymbols / ISessionObjects /
ISessionMemoryAllocator; nothing consumed its members (2 structural
references). Deleted.

IRuntimeSession is the session root (Is64Bit, Memory, Symbols, Objects);
IVBExecutionContext is now just { IRuntimeSession Session } -- the bare
Is64Bit / Memory were unread. Session interfaces move RDCore.Runtime ->
RDCore.SDK/Runtime/Abstract/Execution; impls (RuntimeSession,
SessionSymbols, SessionObjects, SessionMemory, ...) stay in Runtime.
SessionMemoryBlock / SessionMemoryInfo -> RDCore.SDK.Runtime.Shared,
beside MemoryAddress. ISymbolResolver.TryRead(long) ->
TryRead(MemoryAddress). ISymbolProvider kept (planned abstraction for the
config / AST / referenced-library / env-host symbol providers).
