+++ RDCore.Runtime/Execution/Memory/FreeListManager.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using RDCore.SDK.Runtime.Shared;
+using System.Diagnostics.CodeAnalysis;
 
 namespace RDCore.Runtime.Execution.Memory;
