+++ RDCore.SDK/Runtime/Abstract/Execution/ISymbolResolver.cs
@@ -1,0 +1,51 @@
+﻿using RDCore.SDK.Model.Symbols;
+using RDCore.SDK.Model.Symbols.Abstract;
+using RDCore.SDK.Model.Values.Bindings;
+using RDCore.SDK.Runtime.Shared;
+using System.Diagnostics.CodeAnalysis;
+
+namespace RDCore.SDK.Runtime.Abstract.Execution;
+
+/// <summary>
+/// A service that resolves an <em>identifier name</em> to a <c>Symbol</c> given a <em>scope</em> <c>Uri</c>
+/// and that can look up the value currently bound to a symbol (or address) in the current context.
+/// </summary>
+public interface ISymbolResolver
+{
+    /// <summary>
+    /// Resolves the specified <em>identifier name</em> in the specified scope.
+    /// </summary>
+    /// <param name="name">The name of the <see cref="Symbol"/> to resolve.</param>
+    /// <param name="scope">The memory scope to inspect.</param>
+    /// <param name="handle">The <see cref="Uri"/> of the current scope (procedure) symbol.</param>
+    /// <returns><c>null</c> if no symbol could be resolved from the specified <em>handle</em> in the specified <em>scope</em> with the specified <em>name</em>.</returns>
+    Symbol? Resolve(string name, ScopeKind scope, Uri handle);
+
+    /// <summary>
+    /// Gets the <see cref="IBindingHandle"/> currently associated with the specified <see cref="Symbol"/>.
+    /// </summary>
+    /// <param name="symbol">The <see cref="Symbol"/> to retrieve the currently associated binding for.</param>
+    IBindingHandle GetValue(Symbol symbol);
+
+    /// <summary>
+    /// Gets the <see cref="IBindingHandle"/> at the specified address in the runtime <em>memory map</em>.
+    /// </summary>
+    /// <param name="address">The memory address to read.</param>
+    /// <param name="value">The retrieved binding, if successful.</param>
+    bool TryRead(MemoryAddress address, [NotNullWhen(true)][MaybeNullWhen(false)] out IBindingHandle? value);
+}
+
+/// <summary>
+/// A service that loads a <see cref="Symbol"/> into the semantic layer.
+/// </summary>
+/// <remarks>
+/// ⚖️<strong>RDCore</strong> provides an implementation of this interface <strong>licensed under GPLv3</strong>.
+/// </remarks>
+public interface ISymbolProvider
+{
+    /// <summary>
+    /// Defines the specified new <see cref="Symbol"/> in the semantic layer (static context), or in the symbol table (runtime context).
+    /// </summary>
+    /// <param name="symbol">The new <see cref="Symbol"/> to be semantically defined.</param>
+    void Define(Symbol symbol);
+}
