+++ RDCore.SDK/Runtime/Abstract/Execution/IVBExecutionContext.cs
@@ -1,20 +1,12 @@
 ﻿namespace RDCore.SDK.Runtime.Abstract.Execution;
 
 /// <summary>
-/// Represents and encapsulates the execution environment and memory space.
+/// Encapsulates the execution environment for a bound node evaluation.
 /// </summary>
 public interface IVBExecutionContext
 {
     /// <summary>
-    /// <c>true</c> if the execution context describes a 64-bit environment; <c>false</c> otherwise.
+    /// The execution session this evaluation runs in.
     /// </summary>
-    /// <remarks>
-    /// This value determines the size of a <c>VBLongPtrValue</c> and is used anywhere needed, e.g. when evaluating precompiler directives.
-    /// </remarks>
-    bool Is64Bit { get; }
-
-    /// <summary>
-    /// Gets the memory space for this context.
-    /// </summary>
-    IVirtualHeap Memory { get; }
+    IRuntimeSession Session { get; }
 }
